+++ contracts/models/fee/IFeeModel.sol
@@ -6,13 +6,13 @@ interface IFeeModel {
 
     function getInterestFeeAmount(
         address pool,
-        uint256 depositID,
+        uint64 depositID,
         uint256 interestAmount
     ) external view returns (uint256 feeAmount);
 
     function getEarlyWithdrawFeeAmount(
         address pool,
-        uint256 depositID,
+        uint64 depositID,
         uint256 withdrawnDepositAmount
     ) external view returns (uint256 feeAmount);
 }
