+++ contracts/rewards/MPHMinter.sol
@@ -58,7 +58,7 @@ contract MPHMinter is MPHMinterLegacy {
     /**
         v3 functions
      */
-    function createVestForDeposit(address account, uint256 depositID)
+    function createVestForDeposit(address account, uint64 depositID)
         external
         onlyRole(WHITELISTED_POOL_ROLE)
     {
@@ -71,7 +71,7 @@ contract MPHMinter is MPHMinterLegacy {
     }
 
     function updateVestForDeposit(
-        uint256 depositID,
+        uint64 depositID,
         uint256 currentDepositAmount,
         uint256 depositAmount
     ) external onlyRole(WHITELISTED_POOL_ROLE) {
@@ -98,7 +98,7 @@ contract MPHMinter is MPHMinterLegacy {
         return amount;
     }
 
-    function distributeFundingRewards(uint256 fundingID, uint256 interestAmount)
+    function distributeFundingRewards(uint64 fundingID, uint256 interestAmount)
         external
         onlyRole(WHITELISTED_POOL_ROLE)
     {
